ajustes en permisos

+++ resources/views/dashboard/obras/detalle/imagenes-principales/ver.blade.php
@@ -5,7 +5,9 @@
                 <a href="{{ asset('img/obras/imagenes-principales/'.$imagen->imagen_grande) }}" data-gallery="">
                     <img class="img-responsive" src="{{ asset('img/obras/imagenes-principales/'.$imagen->imagen_grande) }}" style="height: 150px; width: auto; margin: auto;">
                 </a>
-                <i onclick="eliminarImagenPrincipalDeObra({{ $imagen->id }}, {{ $imagen->obra_id }})" class="fa fa-trash texto-rojo fa-lg icono-img" aria-hidden="true" data-placement="left" mi-tooltip="Eliminar"></i>
+                @if (Auth::user()->rol->edicion_de_registro_basica || Auth::user()->rol->edicion_de_registro_avanzada_1 || Auth::user()->rol->edicion_de_registro_avanzada_2)
+                    <i onclick="eliminarImagenPrincipalDeObra({{ $imagen->id }}, {{ $imagen->obra_id }})" class="fa fa-trash texto-rojo fa-lg icono-img" aria-hidden="true" data-placement="left" mi-tooltip="Eliminar"></i>
+                @endif
             </div>
         @endforeach
     </div>
